# backend/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    NotificationRead:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: integer
        is_read:
          title: Is Read
          type: boolean
        message:
          title: Message
          type: string
        notification_type:
          title: Notification Type
          type: string
        title:
          title: Title
          type: string
        user_id:
          title: User Id
          type: integer
      required:
      - title
      - message
      - notification_type
      - id
      - user_id
      - is_read
      - created_at
      title: NotificationRead
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login
      type: oauth2
info:
  title: Gyanvruksh API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/dashboard/notifications:
    get:
      description: Get user notifications
      operationId: get_notifications_api_dashboard_notifications_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Notifications
      tags:
      - dashboard
  /api/dashboard/recommendations:
    get:
      description: Get personalized course recommendations
      operationId: get_recommendations_api_dashboard_recommendations_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Recommendations
      tags:
      - dashboard
  /api/dashboard/student/recent-courses:
    get:
      description: Get recently accessed courses for student
      operationId: get_recent_courses_api_dashboard_student_recent_courses_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Recent Courses
      tags:
      - dashboard
  /api/dashboard/student/stats:
    get:
      description: Get dashboard statistics for students
      operationId: get_student_dashboard_stats_api_dashboard_student_stats_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Student Dashboard Stats
      tags:
      - dashboard
  /api/dashboard/teacher/dashboard-stats:
    get:
      description: Get comprehensive dashboard statistics for teachers
      operationId: get_teacher_dashboard_stats_api_dashboard_teacher_dashboard_stats_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Teacher Dashboard Stats
      tags:
      - dashboard
...
  /api/notifications/:
    get:
      operationId: get_notifications_api_notifications__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NotificationRead'
                title: Response Get Notifications Api Notifications  Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Notifications
      tags:
      - notifications
  /api/notifications/{notification_id}/read:
    post:
      operationId: mark_as_read_api_notifications__notification_id__read_post
      parameters:
      - in: path
        name: notification_id
        required: true
        schema:
          title: Notification Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Mark As Read
      tags:
      - notifications
...